Allow setting 'apt update'-specific options
The '--apt-options' setting can only be used for setting options that
are understood by all apt features (install, update, etc.) used by
live-build. Attempting to enable a feature specific to 'apt update',
such as '--error-on=any', will result in a build failure. To allow
these options to be set as well, add an '--apt-update-options' setting
to 'lb config'.